About to pivot the design of the system a bit.
No regions, as this functionality would need to be implemented at the
orchestration level for applications (think k8s), so implementing it
here will only prove to slow down the system and make it a bit more
complicated.

No CRDTs. We need absolute ordering. CRDTs are great when immediate
ordering is not needed and data order just needs to converge at some
point. With railgun, we need immediate ordering.