Initial code for discovery system with DNS backend.
Actix is literally such a great choice for a system like this. 100%,
absolutely, beyond a shadow of a doubt, using actix for this system.

Base line actor setup is good. DNS queries are going through on a
regular interval. This is a solid foundation for adding additional
discovery backends in the future.