Handshake protocol w/ request/respone pattern is g2g.
Request/response system is now in working order in the WsToPeer actor.
This still needs to be replicated in the WsFromPeer actor.

The handshake protocol has been simplified and is fleshed out in the
WsToPeer actor. This needs to be finished up in the WsFromPeer actor.
The gathered information from the handshake needs to be propagated up to
the connections actor and then to the core RG actor.

Connections to self due to discovery protocol will now be dropped with
specific disconnect frames.