Finished routing of outbound frames to target socket.
As the handshake protocol is finished, connections are registering
themselves along with their connected peer's ID with the connections
actor. The combination of the peer ID & the addr of the actor holding
the connection is what comprises the routing table.

Now that the routing table is finished, the connections actor is able to
route outbond request/response frames to the appropriate socket.