Hacking on client routing info exchange.
Baseline models are in place for client routing info exchange is in
place. Work remains to be done, but will be tackled as part of the RPC
implementation.

Refactored storage layer to use AppDataError type, which translates down
to a ClientError as needed, but may also represent control flow patterns
which may still result in an ultimate success. This is needed for
situations where we want an overall success for a client request, but do
not want to write data to Raft.