Storage layer events.
The storage layer event system is in solid shape. These events primarily
flow out to the CPC, and are used to drive placement reconciliation.
Reconciliation code for the CPC is under way.

Lots of refactoring on the database management layer & the CRC database
interaction code. We are now using a system-wide Database object which
manages various aspects of database access and database lifecycle.

Overall, switching to the database manager will allow us to more easily
interact with the database for Streams, Pipelines and other objects
which need to write their data to disk in a predictable location.

CRC system state recovery is nearly finished. Only remaining items are
for users & tokens, which we can knock out once those subsystems are in
a more complete state.

The CRC system index has been updated to use DashMap. This SHOULD give
us less lock contention than the approach being used previously. Lock
contention will still only ever last as long as it takes to clone an
Arc, as that is the only duration of time for which locks are held on
reading.