Stream subscriptions.
Stream subscription controller is handling subscription requests,
recording subscription data & is priming the client subscription channel
for data delivery.

Stream records are stored in an independent tree which enables iteration
and other such behavior directly on the tree without having to perform
manual prefix analysis. Metadata is stored in a separate tree using a
similar canonical naming pattern.

Fetching of data from the stream for subscription delivery is done, and
delivery of data to active subscriber channels is done.