Perform full data reconciliation on leadership change
Operator has been updated to mitigate stale data due to subtleties in
leadership changes and the conditions thereof. In cases where a leader
died and effectively missed reconciliation events, the standby replicas
will not be able to take any action given that they are followers. Once
a node becomes a leader, it will perform a full reconciliation over the
current state of the cluster in order to ensure the system is in the
needed state.

closes #71