Update Pipeline CRD to support starting point config
This update allows pipelines to be created with a configuration which
controls the starting point of the source stream from which the pipeline
should begin processing.

A few efficiency improvements are added as well for encoding pipeline
metadata, storage of data models, and fixes an issue with recording of
active pipeline instances.

closes #68