Using a more robust pattern for JWT creation & verification
Each Token will now have its own HMAC key created for using in the
initial creation of the Token's JWT as well as for all subsequent
verification as the JWT is used.

Downstream systems, such as Streams, simply observe the Secrets and load
their decoding keys for quick and easy verification when establishing
authN & authZ for a new connection to the system.

closes #60