stream: add stream retention policy
Tracking of stream earliest timestamp records is being tracked. This
provides the foundation for our stream compaction/truncation system.

Updated tests to assert the proper functionality of the earliest
timestamp tracking pattern.

Fixed a few small docs items in the pipeline-txp example.

Update pipeline spawning routine to ensure that only pipelines which are
part of the parent stream are spawned.