Add monitoring mixin
Added placeholder values for alerts, rules, and the like. Added a
reference Grafana dashboard along with alerts. Users can utilize this
either programmatically via jsonnet or via manual changes to the
imported dashboard.

Add some docs to the guide on Hadron monitoring, and how to integrate
with the Monitoring Mixins system. Also added some docs to the
Operator's README with info on these capabilities and a link over to the
guide.